+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- 2026-08-05 ButtonGroup/Checkbox: Add tiny size variant
 - 2026-07-11 NumberInput: Support free-form editing with error indication and step-based precision
 - 2026-07-11 NumberInput: Add drag-to-scrub mode with hold-to-activate
 - 2026-07-10 NumberInput: Support custom classes and styles